Revert "Merge branch 'main' into TristinsBingusCave"
This reverts commit 761a1861f38fe8aa84e67dc8d2c618cac3bbae56, reversing
changes made to f2aad26dd802459f2db66d83f76afaea2f58be15.

+++ Assets/Scripts/UI/PlayerHealthBar.cs
@@ -13,7 +13,7 @@ using UnityEngine.UI;
 
 public class PlayerHealthBar : MonoBehaviour
 {
-    private float MaxHealth;
+    private int MaxHealth;
     private PlayerBehaviour playerBehaviour;
     public Slider healthBar;
